removed unneccsary files and added inference-api.py
