+++ Dockerfile.cpu
@@ -1,7 +1,9 @@
+  
 FROM tensorflow/tensorflow:1.12.0-py3
 
 ENV LANG=C.UTF-8
 RUN mkdir /gpt-2
+EXPOSE 5000
 WORKDIR /gpt-2
 ADD . /gpt-2
 RUN pip3 install -r requirements.txt
@@ -9,3 +11,4 @@ RUN python3 download_model.py 124M
 RUN python3 download_model.py 355M
 RUN python3 download_model.py 774M
 RUN python3 download_model.py 1558M
+CMD python3 src/inference-api.py
