+++ Dockerfile.gpu
@@ -11,6 +11,7 @@ ENV NVIDIA_VISIBLE_DEVICES=all \
     LANG=C.UTF-8
 
 RUN mkdir /gpt-2
+EXPOSE 5000
 WORKDIR /gpt-2
 ADD . /gpt-2
 RUN pip3 install -r requirements.txt
@@ -18,3 +19,4 @@ RUN python3 download_model.py 124M
 RUN python3 download_model.py 355M
 RUN python3 download_model.py 774M
 RUN python3 download_model.py 1558M
+CMD python3 src/inference-api.py
