+++ requirements.txt
@@ -2,3 +2,5 @@ fire>=0.1.3
 regex==2017.4.5
 requests==2.21.0
 tqdm==4.31.1
+flask
+flasgger==0.8.3
